feat: pagination improvements
